支持将查询到的备案信息批量写入到Excel表格文件
支持将查询到的备案信息写入到Excel表格文件，Bug fixed

+++ ICP-Checker.py
@@ -56,7 +56,7 @@ def get_cookies():
     return -1
 
 def get_token():
-    timeStamp = int(round(time.time()*1000))
+    timeStamp = round(time.time()*1000)
     authSecret = 'testtest' + str(timeStamp)
     authKey = hashlib.md5(authSecret.encode(encoding='UTF-8')).hexdigest()
     auth_data = {'authKey': authKey,'timeStamp': timeStamp}
